backend/requirements.txt
flask
requests
python-dotenv
deep-translator
transformers
torch